requirements.txt
aerich==0.7.1
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
amqp==5.1.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.0
dictdiffer==0.9.0
factory-boy==3.2.1
Faker==17.0.0
flake8==6.0.0
freezegun==1.2.2
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
isort==5.12.0
kombu==5.2.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
mypy==1.0.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
pypika-tortoise==0.1.6
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-celery==0.0.0
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
